+++ examples/filter/src/app.js
@@ -8,10 +8,6 @@ import {defaultMapStyle, highlightLayerIndex} from './map-style.js';
 
 const token = process.env.MapboxAccessToken; // eslint-disable-line
 
-if (!token) {
-  throw new Error('Please specify a valid mapbox token');
-}
-
 export default class App extends Component {
 
   state = {