+++ examples/layers/src/app.js
@@ -6,10 +6,6 @@ import ControlPanel from './control-panel';
 
 const token = process.env.MapboxAccessToken; // eslint-disable-line
 
-if (!token) {
-  throw new Error('Please specify a valid mapbox token');
-}
-
 export default class App extends Component {
 
   state = {