+++ docs/whats-new.md
@@ -5,7 +5,7 @@ Release date: TBD, target late 2017
 ## Highlights
 
 - **Support for Map Reuse**: A new property `reuseMaps` is provided for applications that create and destroy maps, to help work around a mapbox-gl resource leak issue that can lead to a browser crash in certain situations.
-
+- **No Token warning**: react-map-gl now renders an HTML message if no mapbox token is supplied.
 
 # react-map-gl v3.0